// frontend/generated/SettingsPatch.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SettingsPatch = { silentRun: boolean | null, globalSensitivity: number | null, defaultThresholdMs: number | null, longHoldBypassMs: number | null, startupDelayMs: number | null, processGameModeEnabled: boolean | null, gameModeThresholdMs: number | null, gameModeLongHoldBypassMs: number | null, };